requirements.txt
pillow
PySide6